move about and user profile
