+++ src/app/_@features/about/about.component.scss
